Move backstory summarization to it's own capability package. Refactor to clean up wildcard imports and general application structure. Provide overview of application structure in the README.
