+++ src/main/kotlin/ca/kittle/capabilities/connection/CreateConnections.kt
@@ -1,0 +1,3 @@
+package ca.kittle.capabilities.connection
+
+class CreateConnections
