+++ src/main/kotlin/ca/kittle/capabilities/party/ShowParty.kt
@@ -1,10 +1,10 @@
-package ca.kittle.party
+package ca.kittle.capabilities.party
 
+import ca.kittle.capabilities.party.components.populateParty
+import ca.kittle.capabilities.party.models.Party
+import ca.kittle.capabilities.pc.models.PlayerCharacter
 import ca.kittle.db.models.PartyEntity
 import ca.kittle.db.models.PlayerCharacterEntity
-import ca.kittle.party.components.populateParty
-import ca.kittle.party.models.Party
-import ca.kittle.pc.models.PlayerCharacter
 import com.mongodb.client.MongoDatabase
 import com.mongodb.client.model.Filters
 import io.ktor.server.application.call
