+++ src/main/kotlin/ca/kittle/capabilities/party/components/ListOfParties.kt
@@ -1,6 +1,6 @@
-package ca.kittle.party.components
+package ca.kittle.capabilities.party.components
 
-import ca.kittle.party.models.Party
+import ca.kittle.capabilities.party.models.Party
 import kotlinx.html.FlowContent
 import kotlinx.html.li
 import kotlinx.html.ul
