--- src/main/kotlin/ca/kittle/connection/CreateConnections.kt
@@ -1,4 +1,0 @@
-package ca.kittle.connection
-
-class CreateConnections {
-}
